frontendDevServer/requirements.txt
aioredis==1.2.0
asgiref==3.2.1
asn1crypto==0.24.0
astroid==2.2.5
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.8.1
Automat==0.7.0
autopep8==1.4.4
certifi==2019.6.16
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
colorama==0.4.1
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
Django==2.2.1
django-cors-headers==3.1.0
djangorestframework==3.10.2
Faker==2.0.1
feedparser==5.2.1
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
msgpack==0.6.1
pycodestyle==2.5.0
pycparser==2.19
PyHamcrest==1.9.0
pylint==2.3.1
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
Twisted==19.7.0
txaio==18.8.1
typed-ast==1.3.4
wincertstore==0.2
wrapt==1.11.2
ws4py==0.5.1
zope.interface==4.6.0
